delete old_updater only when update doesn't exist, otherwise rename it to updater
The latest release doesn't including updater due to unexpected fail
while compiling updater with batch script (and not monitored). This code
will preserve updater if the new release doesn't come with new updater.

+++ mangadownloader/forms/frmMain.pas
@@ -1084,7 +1084,12 @@ begin
 
   // remove old updater
   if FileExistsUTF8(FMD_DIRECTORY + 'old_updater.exe') then
-    DeleteFileUTF8(FMD_DIRECTORY + 'old_updater.exe');
+  begin
+    if FileExistsUTF8(FMD_DIRECTORY + 'updater.exe') then
+      DeleteFileUTF8(FMD_DIRECTORY + 'old_updater.exe')
+    else
+      RenameFileUTF8(FMD_DIRECTORY + 'old_updater.exe', FMD_DIRECTORY + 'updater.exe');
+  end;
 
   // TrayIcon
   TrayIcon.Icon.Assign(Application.Icon);
